moved input validation to utility functions
